moves v3 controllers to root level and delete three folder
