+++ app/Http/Controllers/TagsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rogue\Http\Controllers\Three;
+namespace Rogue\Http\Controllers;
 
 use Rogue\Models\Post;
 use Illuminate\Http\Request;
